# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop API",
        "description": "API routes for the Petshop coding exercies",
        "contact": {
            "email": "chris.idakwo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9900",
            "description": "Base API URL"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders",
                "description": "List all orders",
                "operationId": "orders/index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/order/create": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "description": "Create a new order",
                "operationId": "order/create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "order_status_uuid",
                                    "payment_uuid",
                                    "products",
                                    "address"
                                ],
                                "properties": {
                                    "order_status_uuid": {
                                        "description": "Order status UUID",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "payment_uuid": {
                                        "description": "Payment UUID",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "products": {
                                        "description": "Array of objects with product uuid and quantity",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "uuid": {
                                                    "description": "Product UUID",
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "quantity": {
                                                    "description": "Order quantity for product",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object",
                                            "minItems": 1
                                        }
                                    },
                                    "address": {
                                        "description": "Billing and Shipping address",
                                        "properties": {
                                            "billing": {
                                                "description": "Order billing address",
                                                "type": "string"
                                            },
                                            "shipping": {
                                                "description": "Order shipping address",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/order/{uuid}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Fetch an order",
                "description": "Fetch an order",
                "operationId": "order/show",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an existing order",
                "description": "Update an existing order",
                "operationId": "order/update",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "order_status_uuid",
                                    "payment_uuid",
                                    "products",
                                    "address"
                                ],
                                "properties": {
                                    "order_status_uuid": {
                                        "description": "Order status UUID",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "payment_uuid": {
                                        "description": "Payment UUID",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "products": {
                                        "description": "Array of objects with product uuid and quantity",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "uuid": {
                                                    "description": "Product UUID",
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "quantity": {
                                                    "description": "Order quantity for product",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object",
                                            "minItems": 1
                                        }
                                    },
                                    "address": {
                                        "description": "Billing and Shipping address",
                                        "properties": {
                                            "billing": {
                                                "description": "Order billing address",
                                                "type": "string"
                                            },
                                            "shipping": {
                                                "description": "Order shipping address",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an existing order",
                "description": "Delete an existing order",
                "operationId": "order/delete",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/order/{uuid}/download": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Download an order",
                "description": "Download an order",
                "operationId": "order/download",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT authentication token",
                "scheme": "bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}